Refactor submission processing, tests
