+++ app/core/tests/test_core_mailchimp_service.py
