+++ app/core/tests/test_core_slack_service.py
